Copy .utoprc in setup-misc.sh (not setup-opam.sh) so that 3110-tools is already present
A command to copy a .utoprc file from the tools directory was added in
af4b269d5f0f0dd61f776e3c5c203fa892f4d227. However, because the command was
placed in setup-opam.sh, this caused an error: the 3110-tools directory is
cloned in setup-3110.sh, which must occur *after* setup-opam as ocaml, opam,
and ocaml libraries must be available to compile 3110 tools. As a result, the
cp command failed because 3110-tools hadn't been cloned yet.

This moves that cp command to setup-misc.sh to cause the utoprc to be copied *after*
3110-tools has been cloned.